Split core.flow.js into 2 files, fix errors, use new types. Change interfaces to types since it works better. Split due to https://github.com/facebook/flow/issues/7810
